requirements.txt
tensorflow==2.15.0
streamlit==1.29.0
numpy==1.24.3
pillow==10.1.0
opencv-python-headless==4.8.1.78
matplotlib==3.8.2
scikit-learn==1.3.2